issue #18: change okr app directory
