+++ frontend/src/actions/constants.js
@@ -9,12 +9,4 @@ export const API_MESSAGES = {
   // REQUEST
   REQUEST_SEND_REVIEW: 'The review has updated!',
   REQUEST_SEND_REVIEW_ERROR: 'Saving review failed!',
-
-  // OKR
-  ON_SAVE_OKR: 'Your objectives and key result (OKR) has updated successfully!',
-  ON_SAVE_OKR_ERROR: 'Updating objectives and key result (OKR) was failed!',
-  OKR_FETCH_ONE_ERROR: 'OKR error',
-  ON_CREATE_OKR: "You've created OKR successfully!",
-  ON_CREATE_OKR_ERROR: "Error",
-  ON_DELETE_OKR: "You've deleted OKR successfully!",
 }
