+++ frontend/src/actions/index.js
@@ -9,7 +9,7 @@ import ActionTypes from './types';
 import { API_MESSAGES } from './constants';
 
 
-const callAPI = (axiosRequest, actionType, successMessage = null, errorMessage = null) => {
+export const callAPI = (axiosRequest, actionType, successMessage = null, errorMessage = null) => {
   return async (dispatch) => {
     try {
       const payload = (await axiosRequest).data;
@@ -143,51 +143,6 @@ export const configFetchGradeOptions = () => {
   )
 }
 /** END CONFIG ACTIONS */
-
-
-/** START OKR ACTIONS */
-export const okrFetchAll = () => {
-  return callAPI(
-    axios.get(urls.OKR_FETCH_ALL),
-    ActionTypes.OKR_FETCH_ALL,
-  )
-}
-
-export const okrFetchOne = (id) => {
-  return callAPI(
-    axios.get(urls.OKR_FETCH_ONE(id)),
-    ActionTypes.OKR_FETCH_ONE,
-    null,
-    API_MESSAGES.OKR_FETCH_ONE_ERROR,
-  )
-}
-
-export const onSaveOKR = async (id, okr) => {
-  return callAPI(
-    await axios.patch(urls.OKR_DETAIL_URL(id), { ...okr }),
-    ActionTypes.OKR_FETCH_ONE,
-    API_MESSAGES.ON_SAVE_OKR,
-    API_MESSAGES.ON_SAVE_OKR_ERROR,
-  )
-}
-
-export const onCreateOKR = async (okr) => {
-  return callAPI(
-    await axios.post(urls.OKR_DETAIL_URL(), { ...okr }),
-    ActionTypes.OKR_FETCH_ONE,
-    API_MESSAGES.ON_CREATE_OKR,
-    API_MESSAGES.ON_CREATE_OKR_ERROR,
-  )
-}
-
-export const onDeleteOKR = async (id) => {
-  return callAPI(
-    await axios.delete(urls.OKR_DETAIL_URL(id)),
-    null,
-    API_MESSAGES.ON_DELETE_OKR,
-  )
-}
-/** END OKR ACTIONS */
 
 /** START ACCOUNT ACTIONS */
 export const accountFetchAll = () => {
