+++ frontend/src/actions/urls.js
@@ -2,10 +2,6 @@ const RequestURL = (action = '') => (id = '') => (
   id? (action? `/api/requests/${id}/${action}/`:`/api/requests/${id}/`):'/api/requests/'
 )
 
-const OKRURL = (action = '') => (id = '') => (
-  id? (action? `/api/okrs/${id}/${action}/`:`/api/okrs/${id}/`):'/api/okrs/'
-)
-
 const UserURL = (action = '') => (id = '') => (
   id? (action? `/api/account/${id}/${action}/`:`/api/account/${id}/`):'/api/account/'
 )
@@ -24,11 +20,6 @@ const URLS = {
   REQUEST_FETCH_ALL: RequestURL()(),
   REQUEST_FETCH_ONE: RequestURL(),
   REQUEST_SEND_REVIEW: RequestURL(),
-
-  // OKR
-  OKR_FETCH_ALL: OKRURL()(),
-  OKR_FETCH_ONE: OKRURL(),
-  OKR_DETAIL_URL: OKRURL(),
 
   // CONFIG
   CONFIG_FETCH_GRADE_OPTIONS: ConfigURL('grade_options'),
