+++ frontend/src/okr/actions/constants.js
@@ -1,0 +1,8 @@
+export const API_MESSAGES = {
+  ON_SAVE_OKR: 'Your objectives and key result (OKR) has updated successfully!',
+  ON_SAVE_OKR_ERROR: 'Updating objectives and key result (OKR) was failed!',
+  OKR_FETCH_ONE_ERROR: 'OKR error',
+  ON_CREATE_OKR: "You've created OKR successfully!",
+  ON_CREATE_OKR_ERROR: "Error",
+  ON_DELETE_OKR: "You've deleted OKR successfully!",
+}
